requirements/dev.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile dev.in
#
annotated-types==0.7.0
    # via pydantic
anyio==4.8.0
    # via
    #   httpx
    #   jupyter-server
    #   starlette
appnope==0.1.4
    # via ipykernel
argon2-cffi==23.1.0
    # via jupyter-server
argon2-cffi-bindings==21.2.0
    # via argon2-cffi
arrow==1.3.0
    # via isoduration
asttokens==3.0.0
    # via stack-data
async-lru==2.0.4
    # via jupyterlab
attrs==24.3.0
    # via
    #   hypothesis
    #   jsonschema
    #   referencing
autopep8==2.3.2
    # via nbqa
babel==2.16.0
    # via jupyterlab-server
bandit==1.8.2
    # via -r dev.in
beautifulsoup4==4.12.3
    # via nbconvert
bleach[css]==6.2.0
    # via nbconvert
certifi==2024.12.14
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==1.17.1
    # via argon2-cffi-bindings
charset-normalizer==3.4.1
    # via requests
click==8.1.8
    # via uvicorn
comm==0.2.2
    # via
    #   ipykernel
    #   ipywidgets
coverage==7.6.10
    # via -r dev.in
debugpy==1.8.12
    # via ipykernel
decorator==5.1.1
    # via ipython
defusedxml==0.7.1
    # via nbconvert
executing==2.1.0
    # via stack-data
fastapi==0.115.6
    # via -r prod.in
fastjsonschema==2.21.1
    # via nbformat
flake8==7.1.1
    # via -r dev.in
fqdn==1.5.1
    # via jsonschema
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.7
    # via httpx
httpx==0.28.1
    # via
    #   -r dev.in
    #   -r prod.in
    #   jupyterlab
hypothesis==6.124.1
    # via -r dev.in
idna==3.10
    # via
    #   anyio
    #   httpx
    #   jsonschema
    #   requests
iniconfig==2.0.0
    # via pytest
ipykernel==6.29.5
    # via
    #   jupyter
    #   jupyter-console
    #   jupyterlab
ipython==8.31.0
    # via
    #   ipykernel
    #   ipywidgets
    #   jupyter-console
    #   nbqa
ipywidgets==8.1.5
    # via jupyter
isoduration==20.11.0
    # via jsonschema
jedi==0.19.2
    # via ipython
jinja2==3.1.5
    # via
    #   -r prod.in
    #   jupyter-server
    #   jupyterlab
    #   jupyterlab-server
    #   nbconvert
json5==0.10.0
    # via jupyterlab-server
jsonpointer==3.0.0
    # via jsonschema
jsonschema[format-nongpl]==4.23.0
    # via
    #   jupyter-events
    #   jupyterlab-server
    #   nbformat
    #   openapi-schema-validator
    #   openapi-spec-validator
jsonschema-path==0.3.3
    # via openapi-spec-validator
jsonschema-specifications==2024.10.1
    # via
    #   jsonschema
    #   openapi-schema-validator
jupyter==1.1.1
    # via -r dev.in
jupyter-client==8.6.3
    # via
    #   ipykernel
    #   jupyter-console
    #   jupyter-server
    #   nbclient
jupyter-console==6.6.3
    # via jupyter
jupyter-core==5.7.2
    # via
    #   ipykernel
    #   jupyter-client
    #   jupyter-console
    #   jupyter-server
    #   jupyterlab
    #   nbclient
    #   nbconvert
    #   nbformat
jupyter-events==0.11.0
    # via jupyter-server
jupyter-lsp==2.2.5
    # via jupyterlab
jupyter-server==2.15.0
    # via
    #   jupyter-lsp
    #   jupyterlab
    #   jupyterlab-server
    #   notebook
    #   notebook-shim
jupyter-server-terminals==0.5.3
    # via jupyter-server
jupyterlab==4.3.4
    # via
    #   jupyter
    #   notebook
jupyterlab-pygments==0.3.0
    # via nbconvert
jupyterlab-server==2.27.3
    # via
    #   jupyterlab
    #   notebook
jupyterlab-widgets==3.0.13
    # via ipywidgets
lazy-object-proxy==1.10.0
    # via openapi-spec-validator
markdown-it-py==3.0.0
    # via rich
markupsafe==3.0.2
    # via
    #   jinja2
    #   nbconvert
matplotlib-inline==0.1.7
    # via
    #   ipykernel
    #   ipython
mccabe==0.7.0
    # via flake8
mdurl==0.1.2
    # via markdown-it-py
mistune==3.1.0
    # via nbconvert
mypy==1.14.1
    # via -r dev.in
mypy-extensions==1.0.0
    # via
    #   -r dev.in
    #   mypy
nbclient==0.10.2
    # via nbconvert
nbconvert==7.16.5
    # via
    #   jupyter
    #   jupyter-server
nbformat==5.10.4
    # via
    #   jupyter-server
    #   nbclient
    #   nbconvert
nbqa==1.9.1
    # via -r dev.in
nest-asyncio==1.6.0
    # via ipykernel
notebook==7.3.2
    # via
    #   -r dev.in
    #   jupyter
notebook-shim==0.2.4
    # via
    #   jupyterlab
    #   notebook
openapi-schema-validator==0.6.3
    # via openapi-spec-validator
openapi-spec-validator==0.7.1
    # via -r dev.in
overrides==7.7.0
    # via jupyter-server
packaging==24.2
    # via
    #   ipykernel
    #   jupyter-server
    #   jupyterlab
    #   jupyterlab-server
    #   nbconvert
    #   pytest
pandocfilters==1.5.1
    # via nbconvert
parso==0.8.4
    # via jedi
pathable==0.4.4
    # via jsonschema-path
pbr==6.1.0
    # via stevedore
pexpect==4.9.0
    # via ipython
platformdirs==4.3.6
    # via jupyter-core
pluggy==1.5.0
    # via pytest
prometheus-client==0.21.1
    # via jupyter-server
prompt-toolkit==3.0.50
    # via
    #   ipython
    #   jupyter-console
psutil==6.1.1
    # via ipykernel
ptyprocess==0.7.0
    # via
    #   pexpect
    #   terminado
pure-eval==0.2.3
    # via stack-data
pycodestyle==2.12.1
    # via
    #   autopep8
    #   flake8
pycparser==2.22
    # via cffi
pydantic==2.10.5
    # via
    #   -r dev.in
    #   fastapi
pydantic-core==2.27.2
    # via pydantic
pyflakes==3.2.0
    # via flake8
pygments==2.19.1
    # via
    #   ipython
    #   jupyter-console
    #   nbconvert
    #   rich
pytest==8.3.4
    # via -r dev.in
python-dateutil==2.9.0.post0
    # via
    #   arrow
    #   jupyter-client
python-dotenv==1.0.1
    # via -r prod.in
python-json-logger==3.2.1
    # via jupyter-events
pyyaml==6.0.2
    # via
    #   -r prod.in
    #   bandit
    #   jsonschema-path
    #   jupyter-events
pyzmq==26.2.0
    # via
    #   ipykernel
    #   jupyter-client
    #   jupyter-console
    #   jupyter-server
referencing==0.35.1
    # via
    #   jsonschema
    #   jsonschema-path
    #   jsonschema-specifications
    #   jupyter-events
requests==2.32.3
    # via
    #   jsonschema-path
    #   jupyterlab-server
rfc3339-validator==0.1.4
    # via
    #   jsonschema
    #   jupyter-events
    #   openapi-schema-validator
rfc3986-validator==0.1.1
    # via
    #   jsonschema
    #   jupyter-events
rich==13.9.4
    # via bandit
rpds-py==0.22.3
    # via
    #   jsonschema
    #   referencing
send2trash==1.8.3
    # via jupyter-server
six==1.17.0
    # via
    #   python-dateutil
    #   rfc3339-validator
sniffio==1.3.1
    # via anyio
sortedcontainers==2.4.0
    # via hypothesis
soupsieve==2.6
    # via beautifulsoup4
stack-data==0.6.3
    # via ipython
starlette==0.41.3
    # via fastapi
stevedore==5.4.0
    # via bandit
terminado==0.18.1
    # via
    #   jupyter-server
    #   jupyter-server-terminals
tinycss2==1.4.0
    # via bleach
tokenize-rt==6.1.0
    # via nbqa
tomli==2.2.1
    # via nbqa
tornado==6.4.2
    # via
    #   ipykernel
    #   jupyter-client
    #   jupyter-server
    #   jupyterlab
    #   notebook
    #   terminado
traitlets==5.14.3
    # via
    #   comm
    #   ipykernel
    #   ipython
    #   ipywidgets
    #   jupyter-client
    #   jupyter-console
    #   jupyter-core
    #   jupyter-events
    #   jupyter-server
    #   jupyterlab
    #   matplotlib-inline
    #   nbclient
    #   nbconvert
    #   nbformat
types-python-dateutil==2.9.0.20241206
    # via arrow
types-pyyaml==6.0.12.20241230
    # via -r dev.in
typing-extensions==4.12.2
    # via
    #   anyio
    #   fastapi
    #   ipython
    #   mypy
    #   pydantic
    #   pydantic-core
uri-template==1.3.0
    # via jsonschema
urllib3==2.3.0
    # via requests
uvicorn==0.34.0
    # via -r prod.in
wcwidth==0.2.13
    # via prompt-toolkit
webcolors==24.11.1
    # via jsonschema
webencodings==0.5.1
    # via
    #   bleach
    #   tinycss2
websocket-client==1.8.0
    # via jupyter-server
widgetsnbextension==4.0.13
    # via ipywidgets

# The following packages are considered to be unsafe in a requirements file:
# setuptools
